updated opt methods
